pyproject.toml
[project]
name = "edge-tts-for-legado"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=7.2.7",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
]
